Rename review/store.pure.ts to review/state.ts
